requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
streamlit==1.28.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
tiktoken>=0.5.2,<0.6.0
deepgram-sdk 